+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## TBD
+
+### Bug Fixes
+
+- Fix Android context being overwritten by Activity
+[#977](https://github.com/bugsnag/bugsnag-unity/pull/977)
+
 ## 8.9.1 (2026-04-01)
 
 ### Bug Fixes